+++ plugin/plugin_test.go
@@ -4,7 +4,7 @@ import (
 	"context"
 	"testing"
 
-	"github.com/apache/arrow/go/v16/arrow"
+	"github.com/apache/arrow/go/v17/arrow"
 	"github.com/cloudquery/plugin-sdk/v4/message"
 	"github.com/cloudquery/plugin-sdk/v4/schema"
 	"github.com/rs/zerolog"
@@ -54,6 +54,9 @@ func (c *testPluginClient) Write(_ context.Context, res <-chan message.WriteMess
 	return nil
 }
 func (*testPluginClient) Close(context.Context) error {
+	return nil
+}
+func (*testPluginClient) Transform(context.Context, <-chan arrow.Record, chan<- arrow.Record) error {
 	return nil
 }