+++ scalar/scalar.go
@@ -4,10 +4,10 @@ import (
 	"fmt"
 	"reflect"
 
-	"github.com/apache/arrow/go/v16/arrow"
-	"github.com/apache/arrow/go/v16/arrow/array"
-	"github.com/apache/arrow/go/v16/arrow/float16"
-	"github.com/apache/arrow/go/v16/arrow/memory"
+	"github.com/apache/arrow/go/v17/arrow"
+	"github.com/apache/arrow/go/v17/arrow/array"
+	"github.com/apache/arrow/go/v17/arrow/float16"
+	"github.com/apache/arrow/go/v17/arrow/memory"
 	"github.com/cloudquery/plugin-sdk/v4/types"
 	"golang.org/x/exp/maps"
 )